Refactor imgui_ext
